+++ Actividad_Medio_Termino/Assets/Scripts/Framework/Controllers/Laser.cs
@@ -41,31 +41,39 @@ public class Laser : MonoBehaviour
         Destroy(gameObject);
     }
 
-    /*
-    /// <summary>
-    /// OnCollisionEnter maneja la logica de las colisiones de laser
-    /// </summary>
-    private void OnCollisionEnter(Collision collision)
-    {
-        // Ignorar colisiones con otros proyectiles, misiles y laseres
-        if (collision.gameObject.CompareTag("Projectile") || 
-            collision.gameObject.CompareTag("Laser") || 
-            collision.gameObject.CompareTag("Missile"))
-        {
-            return;
-        }
-
-        Debug.Log("Proyectil impactó en el objetivo: " + collision.gameObject.name);
-        // Si choca con un enemigo, destruir proyectil y enemigo
-        Destroy(gameObject);
-    }
-    */
-
     /// <summary>
     /// Configura la direccion del laser
     /// </summary>
     public void SetDirection(Vector3 dir)
     {
         direction = dir;
+    }
+
+    /// <summary>
+    /// OnTriggerEnter maneja la logica de las colisiones del proyectil de enemigos y jefes
+    /// </summary>
+    private void OnTriggerEnter(Collider other)
+    {
+        // Ignorar colisiones con otros proyectiles, misiles y laseres
+        if (other.gameObject.CompareTag("Projectile") || 
+            other.gameObject.CompareTag("Laser") || 
+            other.gameObject.CompareTag("Missile") ||
+            other.gameObject.CompareTag("Player"))
+        {
+            return;
+        }
+
+        // Si el proyectil impacta contra un jugador
+        if(other.gameObject.CompareTag("Enemy") || other.gameObject.CompareTag("Boss"))
+        {
+            Health enemyHealth = other.gameObject.GetComponent<Health>();
+            if(enemyHealth != null)
+            {
+                enemyHealth.TakeDamage(1);
+            }
+        }
+        Debug.Log("Proyectil impactó en el objetivo: " + other.gameObject.name);
+        // Destruir proyectil al colisionar
+        Destroy(gameObject);
     }
 }